Rename dungeon size clamping options and rewrite their descriptions
The existing descriptions don't appear to be entirely accurate in the current state of the game and LethalLevelLoader.

The dungeon size restriction should no longer be necessary with the introduction of the interior's tile size value, which scales the interior's size multiplier based on the average size of tiles in it.

The size restriction's description also used examples for the lerp factor value that appear to be reversed. In the interest of making the functionality as clear as possible the descriptions are rewritten.