Make CullFactory aware that the night vision lights may be visible
If a camera that doesn't activate night vision is rendered last in a pass for one frame, then that will cause the night vision lights to be disabled. Then, when CullFactory culls for the pass in the next frame, it will see that they are disabled, and cull the lights. Thus, they will never be visible.

By setting them visible before the end of every frame, CullFactory will then consider them to be possibly visible and do the necessary visibility testing to determine if they should actually render.