Rework handheld TV explosion code a little to ensure they are removed
DestroyItemInSlotAndSync will now only be called on the owner of the item.

If it is not held, it will use NetworkObject.Despawn() instead of Destroy() to ensure that all clients have the item destroyed.

The actual explosion is triggered in OnDestroy(), which ensures that if the object is destroyed for any reason, it will explode. This should ensure the explosion goes off even if the object is destroyed by the server before the client has the chance to trigger the explosion.