Display animations based on the healing station capacity
The capacity animation is interpolated on a delay based on the update interval. For local clients healing, it is every healing tick, but for other clients it's every quarter second, currently.

When the healing capacity is depleted, the shuttering animation will play, changing the screen to a red one and closing the shutter over it after a delay.